Add ipc functionality for attaching

+++ sdb/src/process.cpp
@@ -17,7 +17,7 @@ namespace
     }
 }
 
-std::unique_ptr<sdb::process> sdb::process::launch(std::filesystem::path path)
+std::unique_ptr<sdb::process> sdb::process::launch(std::filesystem::path path, bool debug)
 {
     pipe channel(true);
     pid_t pid;
@@ -29,7 +29,7 @@ std::unique_ptr<sdb::process> sdb::process::launch(std::filesystem::path path)
     if (pid == 0)
     {
         channel.close_read();
-        if (ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0)
+        if (debug and (ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0))
         {
             exit_with_perror(channel, "Tracing failed");
         }
@@ -51,8 +51,11 @@ std::unique_ptr<sdb::process> sdb::process::launch(std::filesystem::path path)
         error::send(std::string(chars, chars + data.size()));
     }
 
-    std::unique_ptr<process> proc(new process(pid, true));
-    proc->wait_on_signal();
+    std::unique_ptr<process> proc(new process(pid, true, debug));
+    if (debug)
+    {
+        proc->wait_on_signal();
+    }
 
     return proc;
 }
@@ -69,7 +72,7 @@ std::unique_ptr<sdb::process> sdb::process::attach(pid_t pid)
         error::send_errno("Could not attach");
     }
 
-    std::unique_ptr<process> proc(new process(pid, false));
+    std::unique_ptr<process> proc(new process(pid, false, true));
     proc->wait_on_signal();
 
     return proc;
@@ -80,14 +83,17 @@ sdb::process::~process()
     if (pid_ != 0)
     {
         int status;
-        if (state_ == process_state::running)
+        if (is_attached_)
         {
-            kill(pid_, SIGSTOP);
-            waitpid(pid_, &status, 0);
-        }
+            if (state_ == process_state::running)
+            {
+                kill(pid_, SIGSTOP);
+                waitpid(pid_, &status, 0);
+            }
 
-        ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
-        kill(pid_, SIGCONT);
+            ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
+            kill(pid_, SIGCONT);
+        }
 
         if (terminate_on_end_)
         {
